+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/CompileOnlyKitTest.groovy
@@ -1,0 +1,56 @@
+package ru.vyarus.gradle.plugin.pom
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 05.08.2016
+ */
+class CompileOnlyKitTest extends AbstractKitTest {
+
+    def "Check compileOnly compatibility"() {
+        setup:
+        build("""
+            plugins {
+                id 'java'
+                id 'ru.vyarus.pom'
+            }
+
+            group 'ru.vyarus'
+            version 1.0
+            description 'sample description'
+
+            dependencies {
+                compileOnly 'com.google.code.findbugs:annotations:3.0.0'
+                compile 'org.javassist:javassist:3.16.1-GA'
+            }
+
+            publishing {
+                publications {
+                    maven(MavenPublication) {
+                        from components.java
+                    }
+                }
+            }
+
+            model {
+                tasks.generatePomFileForMavenPublication {
+                    destination = file("\$buildDir/generated-pom.xml")
+                }
+            }
+        """)
+
+        when: "run pom task"
+        run('generatePomFileForMavenPublication')
+
+        def pomFile = file("build/generated-pom.xml")
+        def pom = new XmlParser().parse(pomFile)
+        // for debug
+        println pomFile.getText()
+
+        then: "compile dependency scope corrected"
+        pom.dependencies.'*'.find { it.artifactId.text() == 'javassist' }.scope.text() == 'compile'
+
+        then: "compile only does not appear in pom"
+        !pom.dependencies.'*'.find { it.artifactId.text() == 'annotations' }
+
+    }
+}