+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fix properties tag in pom closure
 * Fix non string values support in pom closure
+* Add workaround for tag name clashing with project method names: prefix '_' (prefix removed on xml merging)
 
 ### 1.0.3 (2016-07-29)
 * Fix repeated tags render like developers (#1)